Patch page change and session lost
