+++ webapp/src/App.vue
@@ -7,6 +7,9 @@
           <component :is="Component"/>
         </transition>
       </router-view>
+      <transition>
+        <SessionCountdown v-if="UserStore.player.countDown" :session="UserStore.player.fleet"/>
+      </transition>
     </section>
   </section>
   <Loading :is-loading="false">
@@ -26,13 +29,47 @@ import Loading from "@/vue/templates/Loading.vue";
 import {useI18n} from "vue-i18n";
 import {UserStore} from "@/objects/stores/UserStore.ts";
 import {LocalKey} from "@/objects/stores/LocalStore.ts";
-import {onMounted} from "vue";
+import {onMounted, onUnmounted} from "vue";
 import FirstLogin from "@/vue/templates/FirstLogin.vue";
 import AlertComponent from "@/vue/alert/AlertComponent.vue";
 import {PlayerStates} from "@/objects/Player.ts";
+import {Fleet} from "@/objects/Fleet.ts";
+import {invoke} from "@tauri-apps/api/tauri";
+import {RustSotServer} from "@/objects/SotServer.ts";
+import {Utils} from "@/objects/Utils.ts";
+import SessionCountdown from "@/components/fleet/SessionCountdown.vue";
 
 const {t} = useI18n();
-document.addEventListener('contextmenu', event => event.preventDefault());
+const gameStatusRefresh: number = setInterval(() => {
+  invoke('get_game_object').then((response: any) => {
+    const rustSotServer: RustSotServer = {status: PlayerStates.CLOSED, ip: response.ip, port: response.port}
+    rustSotServer.status = Utils.parseRustPlayerStatus(response.status);
+
+    if (UserStore.player.status != rustSotServer.status) {
+
+      const fleet: Fleet = UserStore.player.fleet as Fleet;
+
+      // Reset player server
+      if (UserStore.player.status == PlayerStates.IN_GAME && rustSotServer.status != PlayerStates.IN_GAME) {
+        fleet.leaveServer();
+      } else if (rustSotServer.ip && UserStore.player.status != PlayerStates.IN_GAME && rustSotServer.status == PlayerStates.IN_GAME) {
+        UserStore.player.server = {
+          connectedPlayers: [],
+          hash: undefined,
+          ip: rustSotServer.ip,
+          location: "",
+          port: rustSotServer.port,
+        }
+        fleet.joinServer();
+      }
+
+      UserStore.player.status = rustSotServer.status;
+      fleet.updateToSession();
+    }
+  })
+}, 400);
+
+//document.addEventListener('contextmenu', event => event.preventDefault());
 onMounted(() => {
   UserStore.init({
     lang: "en",
@@ -48,6 +85,14 @@ window.onbeforeunload = () => {
       JSON.stringify(UserStore.player),
   );
 };
+
+
+onUnmounted(() => {
+  if (UserStore.player.fleet) {
+    UserStore.player.fleet.leaveSession();
+  }
+  clearInterval(gameStatusRefresh)
+});
 </script>
 
 <style scoped lang="scss">
